+++ CHANGELOG.md
@@ -29,11 +29,21 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
-## 2.1.1a2
+## 2.1.1a3
 
-**Oct 25, 2024**
+**Oct 29, 2024**
 ### Added
 - Fix `Catalog::construct_search_parameters` `limit` description in the documentation.
+
+## 2.1.1a2
+
+**Oct 28, 2024**
+- Fix test coverage for `Tasking::construct_order_parameters`.
+
+## 2.1.1a1
+
+**Oct 28, 2024**
+- Fix types of `Asset::asset_id` and `Asset::_get_info`.
 
 ## 2.1.0